FeatureFlagsCo.Experiments/dev-requirements.txt
flask
elasticsearch
elasticsearch_dsl
numpy
scipy
pandas
requests
pika
redis
azure-servicebus
opencensus-ext-azure
flake8
autopep8
pylint